# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tu Vacuna PAI — API
  description: >
    Contrato HTTP de la API de Tu Vacuna PAI. Este documento es la AUTORIDAD del
    transporte HTTP (endpoints, status codes, JSON). La semántica de la
    sincronización (comandos, estados, idempotencia, cursor, conflictos) NO se
    redefine aquí: vive en `docs/synchronization/sync-contract.md`, al que este
    documento referencia.
  version: v1
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /catalogs/insurers:
    get:
      tags: [Catalogs]
      summary: Listar aseguradoras en salud (EPS) activas
      description: >
        Catalogo global de EPS. Con `regime` filtra por regimen incluyendo
        siempre las de regimen `AMBOS`. Solo devuelve `is_active = true`.
      parameters:
        - name: regime
          in: query
          required: false
          schema:
            type: string
            enum: [CONTRIBUTIVO, SUBSIDIADO, AMBOS]
      responses:
        '200':
          description: Lista de aseguradoras activas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HealthInsurer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  # ------------------------------------------------------------------- Sync
  # Semántica completa de comandos/estados/cursor en docs/synchronization/sync-contract.md
...
  /catalogs/effective:
    get:
      tags: [Catalogs]
      summary: Catalogo efectivo de la institucion del actor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Catalogo efectivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveCatalog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /catalogs/vaccines:
    get:
      tags: [Catalogs]
      summary: Catalogo global de vacunas activas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de vacunas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vaccine'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catalogs]
      summary: Crear vacuna (CATALOG_GLOBAL_WRITE)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccineRequest'
      responses:
        '201':
          description: Vacuna creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vaccine'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /catalogs/vaccines/{id}:
    put:
      tags: [Catalogs]
      summary: Actualizar vacuna (optimista por version)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccineRequest'
      responses:
        '200':
          description: Vacuna actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vaccine'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Catalogs]
      summary: Desactivar vacuna (optimista por version)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204':
          description: Vacuna desactivada
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /catalogs/vaccines/{id}/options:
    get:
      tags: [Catalogs]
      summary: Opciones globales de una vacuna
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Opciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catalogs]
      summary: Crear opcion global
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '201':
          description: Opcion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /catalogs/vaccines/{vaccineId}/options/{id}:
    put:
      tags: [Catalogs]
      summary: Actualizar opcion global
      security:
        - bearerAuth: []
      parameters:
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '200':
          description: Opcion actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Catalogs]
      summary: Desactivar opcion global
      security:
        - bearerAuth: []
      parameters:
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204':
          description: Opcion desactivada
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /catalogs/vaccines/{id}/templates:
    get:
      tags: [Catalogs]
      summary: Templates de una vacuna
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catalogs]
      summary: Crear template
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '201':
          description: Template creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /catalogs/vaccines/{vaccineId}/templates/{id}:
    put:
      tags: [Catalogs]
      summary: Actualizar template
      security:
        - bearerAuth: []
      parameters:
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '200':
          description: Template actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Option'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Catalogs]
      summary: Desactivar template
      security:
        - bearerAuth: []
      parameters:
        - { name: vaccineId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204':
          description: Template desactivado
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  # ------------------------------------------------------------- Instituciones
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por Supabase Auth. Spring valida sub, usuario activo, permiso y scope.
  responses:
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Autenticado pero sin permiso o scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Conflicto de versión (optimistic locking) o de identidad
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    # ---- Catalogos
    HealthInsurer:
      type: object
      properties:
        id: { type: string, format: uuid }
        nit: { type: string }
        name: { type: string }
        code: { type: string, nullable: true }
        regime: { type: string, enum: [CONTRIBUTIVO, SUBSIDIADO, AMBOS] }
...
    Vaccine:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        category: { type: string }
        maxDoses: { type: integer }
        minAgeMonths: { type: integer, nullable: true }
        maxAgeMonths: { type: integer, nullable: true }
        hasLaboratory: { type: boolean }
        hasLot: { type: boolean }
        hasSyringe: { type: boolean }
        hasSyringeLot: { type: boolean }
        hasDiluent: { type: boolean }
        hasDropper: { type: boolean }
        hasPneumococcalType: { type: boolean }
        hasVialCount: { type: boolean }
        hasObservation: { type: boolean }
        active: { type: boolean }
        version: { type: integer }
    EffectiveCatalog:
      type: object
      properties:
        vaccines:
          type: array
          items:
            type: object
            properties:
              vaccineId: { type: string, format: uuid }
              name: { type: string }
              code: { type: string }
              doses:
                type: array
                items: { $ref: '#/components/schemas/CatalogOptionItem' }
              pneumococcalTypes:
                type: array
                items: { $ref: '#/components/schemas/CatalogOptionItem' }
              operationalOptions:
                type: array
                items: { $ref: '#/components/schemas/CatalogOptionItem' }
    CatalogOptionItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        fieldType: { type: string }
        value: { type: string }
        displayName: { type: string }
        sortOrder: { type: integer }
        isDefault: { type: boolean }
...
    VaccineRequest:
      type: object
      required: [name, code, category, maxDoses]
      properties:
        name: { type: string }
        code: { type: string }
        category: { type: string }
        maxDoses: { type: integer, minimum: 1 }
        minAgeMonths: { type: integer, nullable: true }
        maxAgeMonths: { type: integer, nullable: true }
        hasLaboratory: { type: boolean }
        hasLot: { type: boolean }
        hasSyringe: { type: boolean }
        hasSyringeLot: { type: boolean }
        hasDiluent: { type: boolean }
        hasDropper: { type: boolean }
        hasPneumococcalType: { type: boolean }
        hasVialCount: { type: boolean }
        hasObservation: { type: boolean }
    Option:
      type: object
      properties:
        id: { type: string, format: uuid }
        vaccineId: { type: string, format: uuid }
        institutionId: { type: string, format: uuid, nullable: true }
        fieldType: { type: string }
        value: { type: string }
        displayName: { type: string }
        sortOrder: { type: integer }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        sourceTemplateId: { type: string, format: uuid, nullable: true }
        version: { type: integer, format: int64 }
    OptionRequest:
      type: object
      required: [fieldType, value, displayName]
      properties:
        fieldType: { type: string }
        value: { type: string }
        displayName: { type: string }
        sortOrder: { type: integer }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        version: { type: integer, format: int64 }
...